Build using VSCode and CMake Tools

+++ src/generic/datavgen.cpp
@@ -2563,7 +2563,7 @@ void wxDataViewMainWindow::OnPaint( wxPaintEvent &WXUNUSED(event) )
         for (unsigned int item = item_start; item < item_last; item++)
         {
             const int h = GetLineHeight(item);
-            if ( item == m_owner->GetHighlightLine() )
+            if ( static_cast<int>(item) == m_owner->GetHighlightLine() )
             {
                 dc.DrawRectangle(xRect, cur_line_start, widthRect, h);
                 break;
